Last modifications from desktop, not working yet at all
