--- VIEW/footer.php
@@ -1,18 +1,0 @@
-<!-- FOOTER -->
-<div id="f">
-    <div class="container">
-        <div class="row centered">
-            <a href="http://www.damienvauchel.com" target="_blank"><i class="fa fa-user" aria-hidden="true"></i></a><a href="#" target="_blank"><i class="fa fa-linkedin"></i></a><a href="#" target="_blank"><i class="fa fa-twitter"></i></a><a href="#" target="_blank"><i class="fa fa-facebook"></i></a><a href="#" target="_blank"><i class="fa fa-github"></i></a>
-        </div><!-- row -->
-    </div><!-- container -->
-</div><!-- Footer -->
-
-<?php require "modal_contact_form.php"; ?>
-
-<!-- Bootstrap core JavaScript
-================================================== -->
-<!-- Placed at the end of the document so the pages load faster -->
-<script src="https://code.jquery.com/jquery-1.10.2.min.js"></script>
-<script src="../assets/js/bootstrap.min.js"></script>
-</body>
-</html>
